Fix in-video-quiz questions not resizing when video is resized
Previously, if the video transitioned from fullscreen to regular size
with a question open, or if the window was resized with a question open,
the question would not resize properly (the position, height, and/or width
would be incorrect). This commit checks for changes to the video dimensions
every 100ms when a question is open, and resizes the question if a change
is detected.